Remove a couple obsolete mixable reference in Pages.pm
Fix tables in Fishbone mixerlink.html and remove a mixable reference
in songinfo.

Use newlines in the import plugins.

Allow each playlist type to be removed specifically when the importer
is turned off or disabled.  Import.pm regenerates the list when no new
imports are running.

Fix a bug where importer use/not use settings were undef on install.

Fix a bug where musicmagic mix settings had no effect.

remove use/not use settings from server settings->plugins.

clean up children linklist when plugins are disabled/enabled.

Commit a fix from Dan that optimises out a few extra contentType test
and db commits for http urls that have nothing to do with ongoing db
info, liek the ones we use to query musicmagic.  increased loading
time of musimagic by nearly 10%

change musicmagic getMix to return a reference instead of an array.



git-svn-id: file:///home/awy/mirror/slim/trunk/server@2080 62299810-d8cb-41fd-93b7-d32162e5a4a4

+++ Slim/DataStores/DBI/DBIStore.pm
@@ -652,13 +652,14 @@ sub addExternalPlaylist {
 
 sub clearExternalPlaylists {
 	my $self = shift;
+	my $url = shift;
 
 	$self->{'externalPlaylists'} = [];
 
 	my $playLists = Slim::DataStores::DBI::Track->externalPlaylists();
 
 	while (my $track = $playLists->next()) {
-		$track->delete();
+		$track->delete() if (defined $url ? $track->url() =~ /^$url/ : 1);
 	}
 
 	$self->forceCommit();
